requirements.txt
celery==5.6.0
redis==7.1.0
fastapi==0.124.4
uvicorn==0.38.0
pydantic==2.12.5
pymongo==4.15.5
langchain-google-community==3.0.2
google-api-python-client>2.100.0
langchain-community==0.4.1
google-search-results==2.4.2
langchain-anthropic>=1.0.0
langchain-openai>=1.0.0
openai==2.11.0
pillow==12.0.0
langgraph==1.0.5